find minimum in a list
Find the minimum in a user inputted list. The strategy here is to take the list of numbers, take out the first number and create a list of the remaining numbers. Then, use a loop to circle through the list of remaining numbers, comparing each one to the first one. If any of these numbers are less than the first one, they automatically become the minimum.